move sync topic reader
