+++ internal/topic/topicreaderinternal/raw_topic_reader_stream_mock_test.go
@@ -3,7 +3,7 @@
 //
 // Generated by this command:
 //
-//	mockgen -destination raw_topic_reader_stream_mock_test.go --typed -package topicreaderinternal -write_package_comment=false . RawTopicReaderStream
+//	mockgen -write_generate_directive -destination raw_topic_reader_stream_mock_test.go --typed -package topicreaderinternal -write_package_comment=false . RawTopicReaderStream
 package topicreaderinternal
 
 import (
@@ -12,6 +12,8 @@ import (
 	rawtopicreader "github.com/ydb-platform/ydb-go-sdk/v3/internal/grpcwrapper/rawtopic/rawtopicreader"
 	gomock "go.uber.org/mock/gomock"
 )
+
+//go:generate mockgen -write_generate_directive -destination raw_topic_reader_stream_mock_test.go --typed -package topicreaderinternal -write_package_comment=false . RawTopicReaderStream
 
 // MockRawTopicReaderStream is a mock of RawTopicReaderStream interface.
 type MockRawTopicReaderStream struct {
